+++ package.json
@@ -3,8 +3,9 @@
   "private": true,
   "version": "0.26.1",
   "main": "output/index.js",
+  "type": "module",
   "scripts": {
-    "build": "rollup --config rollup.config.js"
+    "build": "deno run --allow-write=. ./scripts/updateVersion.ts && rollup --config rollup.config.js"
   },
   "repository": {
     "type": "git",
@@ -20,20 +21,23 @@
     "scripts": "dist/index.js"
   },
   "dependencies": {
+    "@dprint/node-plugin-base": "^0.0.2",
     "prettier": "^2.8.8",
     "prettier-plugin-astro": "^0.0.12",
-    "prettier-plugin-svelte": "^2.10.1",
     "prettier-plugin-jsdoc": "^0.3.38",
+    "prettier-plugin-svelte": "^2.10.1",
     "svelte": "^3.47.0",
-    "tslib": "^2.3.1"
+    "tslib": "^2.6.0"
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^21.0.1",
-    "@rollup/plugin-typescript": "^8.3.0",
-    "@types/node": "^17.0.10",
-    "@types/prettier": "^2.4.3",
+    "@rollup/plugin-commonjs": "^25.0.2",
+    "@rollup/plugin-json": "^6.0.0",
+    "@rollup/plugin-node-resolve": "^15.1.0",
+    "@rollup/plugin-typescript": "^11.1.2",
+    "@types/node": "^20.4.1",
+    "@types/prettier": "^2.7.3",
     "copyfiles": "^2.4.1",
-    "rollup": "^2.66.0",
-    "typescript": "^4.5.5"
+    "rollup": "^3.26.2",
+    "typescript": "^5.1.6"
   }
 }
