--- src/messenger/index.ts
@@ -1,2 +1,0 @@
-export { MessagePart } from "./MessagePart";
-export * from "./StdIoMessenger";
