+++ src/version.ts
@@ -1,0 +1,1 @@
+export const VERSION = "0.26.1";
